Updated SZ3

+++ sz3-sys/SZ3/compressor/SZInterpolationCompressor.hpp
@@ -149,7 +149,6 @@ namespace SZ {
 //            writefile("quant.dat", quant_inds.data(), num_elements);
             encoder.preprocess_encode(quant_inds, 0);
             size_t bufferSize = 1.2 * (quantizer.size_est() + encoder.size_est() + sizeof(T) * quant_inds.size());
-            /* printf("bufferSize = %ld\n", bufferSize); */
 
             uchar *buffer = new uchar[bufferSize];
             uchar *buffer_pos = buffer;
@@ -309,14 +308,14 @@ namespace SZ {
         template<uint NN = N>
         typename std::enable_if<NN == 1, double>::type
         block_interpolation(T *data, std::array<size_t, N> begin, std::array<size_t, N> end, const PredictorBehavior pb,
-                            const std::string &interp_func, const int direction, uint stride = 1) {
+                            const std::string &interp_func, const int direction, size_t stride = 1) {
             return block_interpolation_1d(data, begin[0], end[0], stride, interp_func, pb);
         }
 
         template<uint NN = N>
         typename std::enable_if<NN == 2, double>::type
         block_interpolation(T *data, std::array<size_t, N> begin, std::array<size_t, N> end, const PredictorBehavior pb,
-                            const std::string &interp_func, const int direction, uint stride = 1) {
+                            const std::string &interp_func, const int direction, size_t stride = 1) {
             double predict_error = 0;
             size_t stride2x = stride * 2;
             const std::array<int, N> dims = dimension_sequences[direction];
@@ -340,7 +339,7 @@ namespace SZ {
         template<uint NN = N>
         typename std::enable_if<NN == 3, double>::type
         block_interpolation(T *data, std::array<size_t, N> begin, std::array<size_t, N> end, const PredictorBehavior pb,
-                            const std::string &interp_func, const int direction, uint stride = 1) {
+                            const std::string &interp_func, const int direction, size_t stride = 1) {
             double predict_error = 0;
             size_t stride2x = stride * 2;
             const std::array<int, N> dims = dimension_sequences[direction];
@@ -384,7 +383,7 @@ namespace SZ {
         template<uint NN = N>
         typename std::enable_if<NN == 4, double>::type
         block_interpolation(T *data, std::array<size_t, N> begin, std::array<size_t, N> end, const PredictorBehavior pb,
-                            const std::string &interp_func, const int direction, uint stride = 1) {
+                            const std::string &interp_func, const int direction, size_t stride = 1) {
             double predict_error = 0;
             size_t stride2x = stride * 2;
             max_error = 0;
